De-identification Scripts
The scripts will read in the identified data. With NRIC and NAME in the first two columns - two levels of de-identification will take place, applying SHA256 hash on the NRIC and NAME. With NRIC-sha256 being the primary key and NAME-sha256 being the secondary key for subsequent data linkages. The scripts will output a master key file and a de-identified data set for use.